requirements.txt
cached-property==1.3.0
coloredlogs==5.0
Django==1.9
djangorestframework==3.4.0
factory-boy==2.7.0
fake-factory==0.5.9
funcsigs==1.0.2
google-api-python-client==1.5.1
httplib2==0.9.2
humanfriendly==1.44.7
ipaddress==1.0.16
mock==2.0.0
oauth2client==2.2.0
pbr==1.10.0
pip-accel==0.43
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
transaction==1.6.1
uritemplate==0.6
zope.interface==4.2.0
